+++ Makefile-tests.am
@@ -20,39 +20,38 @@ endif
 
 GITIGNOREFILES += ssh-config ansible-inventory.yml vmcheck-logs/ test-compose-logs/ tests/vmcheck/image.qcow2
 
-# Disabled for now because we dropped librpmostreepriv.la
-# testbin_cppflags = $(AM_CPPFLAGS) -I $(srcdir)/src/lib -I $(srcdir)/src/libpriv -I $(srcdir)/libglnx -I $(srcdir)/tests/common
-# testbin_cflags = $(AM_CFLAGS) -fvisibility=hidden $(PKGDEP_RPMOSTREE_CFLAGS)
-# testbin_ldadd = $(PKGDEP_RPMOSTREE_LIBS) librpmostree-1.la librpmostreepriv.la -lstdc++
-# 
-# noinst_LTLIBRARIES += libtest.la
-# libtest_la_SOURCES = tests/common/libtest.c
-# libtest_la_CPPFLAGS = $(testbin_cppflags)
-# libtest_la_CFLAGS = $(testbin_cflags)
-# libtest_la_LIBADD = $(testbin_ldadd)
-# 
-# tests_check_jsonutil_CPPFLAGS = $(testbin_cppflags)
-# tests_check_jsonutil_CFLAGS = $(testbin_cflags)
-# tests_check_jsonutil_LDADD = $(testbin_ldadd) libtest.la
-# 
-# tests_check_postprocess_CPPFLAGS = $(testbin_cppflags)
-# tests_check_postprocess_CFLAGS = $(testbin_cflags)
-# tests_check_postprocess_LDADD = $(testbin_ldadd) libtest.la
-# 
-# tests_check_test_utils_CPPFLAGS = $(testbin_cppflags)
-# tests_check_test_utils_CFLAGS = $(testbin_cflags)
-# tests_check_test_utils_LDADD = $(testbin_ldadd) libtest.la
-# 
-# tests_check_test_sysusers_CPPFLAGS = $(testbin_cppflags)
-# tests_check_test_sysusers_CFLAGS = $(testbin_cflags)
-# tests_check_test_sysusers_LDADD = $(testbin_ldadd) libtest.la
-# 
-# uninstalled_test_programs = \
-# 	tests/check/jsonutil			\
-# 	tests/check/postprocess			\
-# 	tests/check/test-utils			\
-# 	tests/check/test-sysusers		\
-# 	$(NULL)
+testbin_cppflags = $(AM_CPPFLAGS) -I $(srcdir)/src/lib -I $(srcdir)/src/libpriv -I $(srcdir)/libglnx -I $(srcdir)/tests/common
+testbin_cflags = $(AM_CFLAGS) $(rpmostree_bin_common_cflags)
+testbin_ldadd = $(rpmostree_bin_common_libs) librpmostreeinternals.la
+
+noinst_LTLIBRARIES += libtest.la
+libtest_la_SOURCES = tests/common/libtest.c
+libtest_la_CPPFLAGS = $(testbin_cppflags)
+libtest_la_CFLAGS = $(testbin_cflags)
+libtest_la_LIBADD = $(testbin_ldadd)
+
+tests_check_jsonutil_CPPFLAGS = $(testbin_cppflags)
+tests_check_jsonutil_CFLAGS = $(testbin_cflags)
+tests_check_jsonutil_LDADD = $(testbin_ldadd) libtest.la
+
+tests_check_postprocess_CPPFLAGS = $(testbin_cppflags)
+tests_check_postprocess_CFLAGS = $(testbin_cflags)
+tests_check_postprocess_LDADD = $(testbin_ldadd) libtest.la
+
+tests_check_test_utils_CPPFLAGS = $(testbin_cppflags)
+tests_check_test_utils_CFLAGS = $(testbin_cflags)
+tests_check_test_utils_LDADD = $(testbin_ldadd) libtest.la
+
+tests_check_test_sysusers_CPPFLAGS = $(testbin_cppflags)
+tests_check_test_sysusers_CFLAGS = $(testbin_cflags)
+tests_check_test_sysusers_LDADD = $(testbin_ldadd) libtest.la
+
+uninstalled_test_programs = \
+	tests/check/jsonutil			\
+	tests/check/postprocess			\
+	tests/check/test-utils			\
+	tests/check/test-sysusers		\
+	$(NULL)
 
 uninstalled_test_scripts = \
 	tests/check/test-lib-introspection.sh \
