+++ src/app/rpmostreemain.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <glib.h>
+
+G_BEGIN_DECLS
+
+int rpmostree_main (int argc, char **argv);
+
+G_END_DECLS
